update: move some files to template utilities
